New: VerifiedIcon widget
moves logic (that was spread over all the widgets) into a new widget.
This widget can also be used in views of custom themes.
