+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
 - Enh: Improve class names
 - Fix: Enabling from profile
 - Enh: Comments
+- Enh: new VerifiedIcon widget
 
 ### 1.0.0-beta.2 (4/16/2023)
 - Enh: Space configurations
