+++ docs/README.md
@@ -9,3 +9,9 @@ Verify your HumHub users with a verified tick on their profiles.
 [Green Meteor](https://github.com/GreenMeteor)
 
 [Felix Hahn](https://github.com/felixhahnweilheim)
+
+### For theme developpers
+The module provides a widget which returns the verified icon (or an empty string).
+You can use it in your theme views as follows:
+
+`<?= humhub\modules\verified\widgets\VerifiedIcon::widget(['container' => $space]); ?>`
